Three major improvements:
1. As an alternative to an actual tape drive, the tar to files on the
local system using the "TapeDriveEmulator" class.  See "emulator_dir"
config option.

2. Previously the system would not / could not recurse through
sub-directories of a target.  Now it does.

3. The system is able to exclude sub-directories from a target, if
those sub-directories are listed as separate targets.  For this to
work, sub-directory targets must be introduced (imported and scanned)
after the parent directory target.  In fact a re-scan does not take
place; files in the db will simply be re-assigned from the parent to
the child.